fix: define `PYTHONPATH` in script filters
This moves the `PYTHONPATH` definition from a single workflow environment variable to a line in each script filter.
It's not as nice of a solution, but it avoids the issue where env var values are retained from the previous version when a user updates the workflow.
That functionality is optional but is enabled by default and will realistically be desired by users to retain all other settings.